+++ packages/ioc/src/global.ts
@@ -1,11 +1,10 @@
-import * as assert from 'assert';
+import assert from 'assert';
 import { CoreContainer } from './container';
 
 //SIDE EFFECT: GLOBAL DI
 let containerSubject: CoreContainer;
 
 /**
-  * @internal
   * Don't use this unless you know what you're doing. Destroys old containerSubject if it exists and disposes everything
   * then it will swap
   */
@@ -17,7 +16,6 @@ export async function __swap_container(c: CoreContainer) {
 }
 
 /**
-  * @internal
   * Don't use this unless you know what you're doing. Destroys old containerSubject if it exists and disposes everything
   * then it will swap
   */
@@ -25,12 +23,17 @@ export function __add_container(key: string, v: object) {
     containerSubject.addSingleton(key, v);
 }
 
+/**
+  * Initiates the global api.
+  * Once this is finished, the Service api and the other global api is available
+  */
 export function __init_container(options: {
     autowire: boolean;
     path?: string | undefined;
 }) {
     containerSubject = new CoreContainer(options);
 }
+
 /**
  * Returns the underlying data structure holding all dependencies.
  * Exposes methods from iti
@@ -44,4 +47,28 @@ export function useContainerRaw() {
     return containerSubject;
 }
 
-
+/**
+ * The Service api, retrieve from the globally init'ed container
+ * Note: this method only works AFTER your container has been initiated
+ * @since 3.0.0
+ * @example
+ * ```ts
+ * const client = Service('@sern/client');
+ * ```
+ * @param key a key that corresponds to a dependency registered.
+ *
+ */
+export function Service<const T>(key: PropertyKey) {
+    const dep = useContainerRaw().get<T>(key)!;
+    assert(dep, "Requested key " + String(key) + " returned undefined");
+    return dep;
+}
+/**
+ * @since 3.0.0
+ * The plural version of {@link Service}
+ * @returns array of dependencies, in the same order of keys provided
+ */
+export function Services<const T extends string[], V>(...keys: [...T]) {
+    const container = useContainerRaw();
+    return keys.map(k => container.get(k)!) as V;
+}
