+++ packages/ioc/src/index.ts
@@ -1,2 +1,2 @@
-export { Service, Services } from './dependency-injection';
-export { CoreContainer } from './container'
+export { Service, Services, __init_container, __swap_container, __add_container } from './global';
+export { Container } from './container'
